% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_count.R
\name{filter_count}
\alias{filter_count}
\title{Filters data on age group and count by primary_fur_color}
\usage{
filter_count(x, my_age)
}
\arguments{
\item{x}{data.frame with columns named age and primary_fur_color}

\item{my_age}{Character. Age group: Adult or Juvenile}
}
\description{
Filters data on age group and count by primary_fur_color
}
\examples{
library(dplyr)
my_squirrels <- tibble(age = c("Adult", "Juvenile"),
                       primary_fur_color = c("Gray", "Black"))
filter_count(x = my_squirrels, my_age = "Adult")
}
